+++ react-app/src/store/constants.js
@@ -1,0 +1,3 @@
+// Session Constants
+export const SET_SESSION = 'session/SET_SESSION';
+export const REMOVE_SESSION = 'session/REMOVE_SESSION';
